Rename namespaces as agreed
