+++ cartoframes/data/observatory/category.py
@@ -1,7 +1,7 @@
 import pandas as pd
 
-from cartoframes.data.catalog.repository.category_repo import get_category_repo
-from cartoframes.data.catalog.repository.dataset_repo import get_dataset_repo
+from cartoframes.data.observatory.repository.category_repo import get_category_repo
+from cartoframes.data.observatory.repository.dataset_repo import get_dataset_repo
 
 
 _CATEGORY_ID_FIELD = 'id'
