+++ test/data/observatory/repository/test_dataset_repo.py
@@ -1,9 +1,9 @@
 import unittest
 
-from cartoframes.data.catalog.dataset import Datasets
+from cartoframes.data.observatory.dataset import Datasets
 
-from cartoframes.data.catalog.repository.dataset_repo import DatasetRepository
-from cartoframes.data.catalog.repository.repo_client import RepoClient
+from cartoframes.data.observatory.repository.dataset_repo import DatasetRepository
+from cartoframes.data.observatory.repository.repo_client import RepoClient
 from ..examples import test_dataset1, test_datasets
 
 try:
